moved rate limit to a separate module
