+++ chain/network/src/concurrency/mod.rs
@@ -1,6 +1,7 @@
 pub mod arc_mutex;
 pub mod atomic_cell;
 pub mod demux;
+pub mod rate;
 pub mod rayon;
 
 #[cfg(test)]
