+++ chain/network/src/concurrency/tests.rs
@@ -1,5 +1,6 @@
 use crate::concurrency::arc_mutex::ArcMutex;
 use crate::concurrency::demux;
+use crate::concurrency::rate;
 use crate::concurrency::rayon;
 
 // drop a trivial future without completion => panic (in debug mode at least).
@@ -17,7 +18,7 @@ async fn must_complete_ok() {
 
 #[tokio::test]
 async fn test_demux() {
-    let demux = demux::Demux::new(demux::RateLimit { qps: 50., burst: 1 });
+    let demux = demux::Demux::new(rate::Limit { qps: 50., burst: 1 });
     for _ in 0..5 {
         let mut handles = vec![];
         for i in 0..1000 {
@@ -40,7 +41,7 @@ async fn test_demux() {
 fn demux_runtime_dropped_before_call() {
     let r1 = tokio::runtime::Runtime::new().unwrap();
     let r2 = tokio::runtime::Runtime::new().unwrap();
-    let demux = r1.block_on(async { demux::Demux::new(demux::RateLimit { qps: 1., burst: 1000 }) });
+    let demux = r1.block_on(async { demux::Demux::new(rate::Limit { qps: 1., burst: 1000 }) });
     drop(r1);
     let call = demux.call(0, |is: Vec<u64>| async { is });
     assert_eq!(Err(demux::ServiceStoppedError), r2.block_on(call));
@@ -50,7 +51,7 @@ fn demux_runtime_dropped_before_call() {
 fn demux_runtime_dropped_during_call() {
     let r1 = tokio::runtime::Runtime::new().unwrap();
     let r2 = tokio::runtime::Runtime::new().unwrap();
-    let demux = r1.block_on(async { demux::Demux::new(demux::RateLimit { qps: 1., burst: 1000 }) });
+    let demux = r1.block_on(async { demux::Demux::new(rate::Limit { qps: 1., burst: 1000 }) });
 
     // Start the call and pause.
     let (send, recv) = tokio::sync::oneshot::channel();
