+++ chain/network/src/config.rs
@@ -1,5 +1,5 @@
 use crate::blacklist;
-use crate::concurrency::demux;
+use crate::concurrency::rate;
 use crate::network_protocol::PeerAddr;
 use crate::network_protocol::PeerInfo;
 use crate::peer_manager::peer_manager_actor::Event;
@@ -66,9 +66,6 @@ impl ValidatorConfig {
 }
 
 #[derive(Clone)]
-pub struct Features {}
-
-#[derive(Clone)]
 pub struct Tier1 {
     /// Interval between broacasts of the list of validator's proxies.
     /// Before the broadcast, validator tries to establish all the missing connections to proxies.
@@ -132,9 +129,9 @@ pub struct NetworkConfig {
     /// Whether this is an archival node.
     pub archive: bool,
     /// Maximal rate at which SyncAccountsData can be broadcasted.
-    pub accounts_data_broadcast_rate_limit: demux::RateLimit,
+    pub accounts_data_broadcast_rate_limit: rate::Limit,
     /// Maximal rate at which RoutingTableUpdate can be sent out.
-    pub routing_table_update_rate_limit: demux::RateLimit,
+    pub routing_table_update_rate_limit: rate::Limit,
     /// Config of the TIER1 network.
     pub tier1: Option<Tier1>,
 
@@ -254,8 +251,8 @@ impl NetworkConfig {
             push_info_period: time::Duration::milliseconds(100),
             outbound_disabled: false,
             archive,
-            accounts_data_broadcast_rate_limit: demux::RateLimit { qps: 0.1, burst: 1 },
-            routing_table_update_rate_limit: demux::RateLimit { qps: 0.5, burst: 1 },
+            accounts_data_broadcast_rate_limit: rate::Limit { qps: 0.1, burst: 1 },
+            routing_table_update_rate_limit: rate::Limit { qps: 0.5, burst: 1 },
             tier1: Some(Tier1 { advertise_proxies_interval: time::Duration::minutes(15) }),
             inbound_disabled: cfg.experimental.inbound_disabled,
             skip_tombstones: if cfg.experimental.skip_sending_tombstones_seconds > 0 {
@@ -319,8 +316,8 @@ impl NetworkConfig {
             outbound_disabled: false,
             inbound_disabled: false,
             archive: false,
-            accounts_data_broadcast_rate_limit: demux::RateLimit { qps: 100., burst: 1000000 },
-            routing_table_update_rate_limit: demux::RateLimit { qps: 100., burst: 1000000 },
+            accounts_data_broadcast_rate_limit: rate::Limit { qps: 100., burst: 1000000 },
+            routing_table_update_rate_limit: rate::Limit { qps: 100., burst: 1000000 },
             tier1: Some(Tier1 {
                 // Interval is very large, so that it doesn't happen spontaneously in tests.
                 // It should rather be triggered manually in tests.
@@ -364,6 +361,9 @@ impl NetworkConfig {
         self.accounts_data_broadcast_rate_limit
             .validate()
             .context("accounts_Data_broadcast_rate_limit")?;
+        self.routing_table_update_rate_limit
+            .validate()
+            .context("routing_table_update_rate_limit")?;
         Ok(VerifiedConfig { node_id: self.node_id(), inner: self })
     }
 }
